requirements.txt
asgiref==3.8.1
Django==5.2.1
django-ckeditor==6.7.2
django-filter==25.1
django-js-asset==3.1.2
django-stubs==5.2.0
django-stubs-ext==5.2.0
django-tinymce==4.1.0
djangorestframework==3.16.0
Markdown==3.8
pillow==11.2.1
psycopg==3.2.9
psycopg2-binary==2.9.10
sqlparse==0.5.3
types-PyYAML==6.0.12.20250516
typing_extensions==4.13.2
